make Stats still directly available to user and reorganize module
