+++ pytissueoptics/rayscattering/__init__.py
@@ -3,6 +3,7 @@ from .materials import ScatteringMaterial
 from .energyLogging import EnergyLogger
 from .source import PencilPointSource, IsotropicPointSource, DirectionalSource, DivergentSource
 from .scatteringScene import ScatteringScene
+from .statistics import Stats
 from .display.viewer import Viewer, PointCloudStyle, Visibility, ViewGroup, Direction
 from .display.views import View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
     View2DSurface, View2DSurfaceX, View2DSurfaceY, View2DSurfaceZ, View2DSlice, View2DSliceX, View2DSliceY, View2DSliceZ
@@ -12,4 +13,4 @@ __all__ = ["Photon", "ScatteringMaterial", "PencilPointSource", "IsotropicPointS
            "DivergentSource", "EnergyLogger", "ScatteringScene", "Viewer", "PointCloudStyle", "Visibility", "ViewGroup",
            "Direction", "View2DProjection", "View2DProjectionX", "View2DProjectionY", "View2DProjectionZ",
            "View2DSurface", "View2DSurfaceX", "View2DSurfaceY", "View2DSurfaceZ", "View2DSlice", "View2DSliceX",
-           "View2DSliceY", "View2DSliceZ", "samples"]
+           "View2DSliceY", "View2DSliceZ", "samples", "Stats"]
